Refactored I2C code.
I2C code now uses header files.
