+++ i2c_slave.h
@@ -1,0 +1,11 @@
+/*
+*
+*/
+
+void TWI_init_slave(void);
+void TWI_match_read_slave(void);
+void TWI_read_slave(void);
+void TWI_match_write_slave(void);
+void TWI_write_slave(void);
+
+unsigned char write_data,recv_data;
