+++ main.c
@@ -1,121 +1,18 @@
 /*
- * TSEA29.c
- * Program for Master Mode
- * Created: 2015-11-05 13:27:23
- * Author : Viktor
- * SDA = Serial DAta
- * SCL = Serial CLock
+ * com-module.c
+ *
+ * Created: 2015-11-06 10:54:37
+ * Author : Victor
  */ 
-
 #include <avr/io.h>
-#include <util/delay.h>
-#include <inttypes.h>
-
-void TWI_start(void);
-void TWI_repeated_start(void);
-void TWI_init_master(void);
-void TWI_write_address(unsigned char);
-void TWI_read_address(unsigned char);
-void TWI_write_data(unsigned char);
-void TWI_read_data(void);
-void TWI_stop(void);
-
-unsigned char address=0x20, read=1, write=0;
-unsigned char write_data=0x01, recv_data;
+#include <i2c_slave.h>	// Com module is a Slave on the I2C-bus. 
 
 int main(void)
 {
-	_delay_ms(2000);
-	DDRB=0xff;
-	TWI_init_master();  // Function to initialize TWI
-	while(1)
-	{
-		if(write_data==0x00)
-		write_data=1;
-		
-		TWI_start(); // Function to send start condition
-		TWI_write_address(address+write); // Function to write address and data direction bit(write) on SDA
-		TWI_write_data(write_data);      // Function to write data in slave
-		TWI_stop(); // Function to send stop condition
-		
-		
-		_delay_ms(10); // Delay of 10 milliseconds
-		
-		TWI_start();
-		TWI_read_address(address+read); // Function to write address and data direction bit(read) on SDA
-		TWI_read_data(); // Function to read data from slave
-		TWI_stop();
-		
-		_delay_ms(1000);
-		
-		write_data = write_data * 2;
-	}
 	
-	
-}
-
-void TWI_init_master(void) // Function to initialize master
-{
-	TWBR=0x4E;    // Bit rate ---- Bit rate = 0x4E(=78) and Prescale = 64 => SCL = 100KHz
-	TWSR=(1<<TWPS1)|(1<<TWPS0);    // Setting prescalar bits - do we need prescalar bits?
-	// SCL freq= F_CPU/(16+2(TWBR).4^TWPS)
-	
-	// TWCR=(1<<TWEN); // Enable TWI?	
-}
-
-void TWI_start(void)
-{
-	// Clear TWI interrupt flag, Put start condition on SDA, Enable TWI
-	TWCR= (1<<TWINT)|(1<<TWSTA)|(1<<TWEN);
-	while(!(TWCR & (1<<TWINT))); // Wait till start condition is transmitted
-	while((TWSR & 0xF8)!= 0x08); // Check for the acknowledgement
-}
-
-void TWI_repeated_start(void)
-{
-	// Clear TWI interrupt flag, Put start condition on SDA, Enable TWI
-	TWCR= (1<<TWINT)|(1<<TWSTA)|(1<<TWEN);
-	while(!(TWCR & (1<<TWINT))); // wait till restart condition is transmitted
-	while((TWSR & 0xF8)!= 0x10); // Check for the acknowledgement
-}
-
-void TWI_write_address(unsigned char data)
-{
-	TWDR=data; // Address and write instruction
-	TWCR=(1<<TWINT)|(1<<TWEN);    // Clear TWI interrupt flag,Enable TWI
-	while (!(TWCR & (1<<TWINT))); // Wait till complete TWDR byte transmitted
-	while((TWSR & 0xF8)!= 0x18);  // Check for the acknowledgement
-}
-
-void TWI_read_address(unsigned char data)
-{
-	TWDR=data;					  // Address and read instruction
-	TWCR=(1<<TWINT)|(1<<TWEN);    // Clear TWI interrupt flag,Enable TWI
-	while (!(TWCR & (1<<TWINT))); // Wait till complete TWDR byte received
-	while((TWSR & 0xF8)!= 0x40);  // Check for the acknowledgement
-}
-
-void TWI_write_data(unsigned char data)
-{
-	TWDR=data;    // put data in TWDR
-	TWCR=(1<<TWINT)|(1<<TWEN);    // Clear TWI interrupt flag,Enable TWI
-	while (!(TWCR & (1<<TWINT))); // Wait till complete TWDR byte transmitted
-	while((TWSR & 0xF8) != 0x28); // Check for the acknowledgement
-}
-
-void TWI_read_data(void)
-{
-	TWCR=(1<<TWINT)|(1<<TWEN);    // Clear TWI interrupt flag,Enable TWI
-	while (!(TWCR & (1<<TWINT))); // Wait till complete TWDR byte transmitted
-	while((TWSR & 0xF8) != 0x58); // Check for the acknowledgement
-	recv_data=TWDR;
-	PORTB=recv_data;	
-}
-
-void TWI_stop(void)
-{
-	// Clear TWI interrupt flag, Put stop condition on SDA, Enable TWI
-	TWCR= (1<<TWINT)|(1<<TWEN)|(1<<TWSTO);
-	while(!(TWCR & (1<<TWSTO)));  // Wait till stop condition is transmitted
+	while (1) 
+    {
+		// Do nothing
+    }
 }
